Create config files
