+++ src/utils/server/sendgrid/marketing/customFields.ts
@@ -1,25 +1,12 @@
 import { UserActionType } from '@prisma/client'
 import * as Sentry from '@sentry/nextjs'
 
+import {
+  SENDGRID_CUSTOM_FIELDS,
+  SENDGRID_RESERVED_FIELDS,
+  SendgridField,
+} from '@/utils/server/sendgrid/marketing/constants'
 import { SendgridClient } from '@/utils/server/sendgrid/sendgridClient'
-
-/**
- * Sendgrid reserved fields. At least one of email or external_id is required.
- */
-export const SENDGRID_RESERVED_FIELDS = [
-  'external_id',
-  'email',
-  'first_name',
-  'last_name',
-  'address_line_1',
-  'address_line_2',
-  'city',
-  'state_province_region',
-  'country',
-  'postal_code',
-  'phone_number',
-] as const
-export type SendgridReservedField = (typeof SENDGRID_RESERVED_FIELDS)[number]
 
 export type SendgridUserActionCustomField = `${UserActionType}_actions`
 export function getSendgridUserActionCustomFieldName(
@@ -27,21 +14,6 @@ export function getSendgridUserActionCustomFieldName(
 ): SendgridUserActionCustomField {
   return `${actionType}_actions`
 }
-
-/**
- * Add new custom fields here.
- *
- * Note: SendGrid has a limit of 500 custom fields.
- */
-export const SENDGRID_CUSTOM_FIELDS = [
-  'signup_date',
-  'user_actions_count',
-  'session_id',
-  ...Object.values(UserActionType).map(getSendgridUserActionCustomFieldName),
-] as const
-export type SendgridCustomField = (typeof SENDGRID_CUSTOM_FIELDS)[number]
-
-export type SendgridField = SendgridReservedField | SendgridCustomField
 
 /**
  * Sendgrid supported field types.
